requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pandas_ta==0.3.14b0
pillow==11.2.1
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
